[Projects] Fold the Orca leader-proxy into iguazio/v4.py, drop dead follower code
Orca is the IG4/Oris API service - the same client mlrun already talks to
for IGUAZIO_V4 auth/events/authorization, so the standalone orca client
package is redundant. iguazio/v4.py's Client now implements
project_leader.Member (the Orca create/update/delete/get-project proxy,
reusing BaseClient's session/auth-validation/error-mapping) instead of
project_follower.Member, whose 2PC hooks and OPA-policy CRUD methods were
dead scaffolding for a scrapped "mlrun leads in IG4/Oris" feature - Orca
leads and owns OPA itself. Dropped the matching "igz" follower registration
in leader.py and its now-dead tests in test_iguazio_v4.py.
